Actualización de main y documentos por subir

+++ src/main/java/org/uma/mbd/mdJarras/jarras/Jarra.java
@@ -7,11 +7,11 @@ public class Jarra {
         contenido = 0;
     }
 
-    public double getCapacidad() {
+    public int getCapacidad() {
         return capacidad;
     }
 
-    public double getContenido() {
+    public int getContenido() {
         return contenido;
     }
